backend/requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-rest-auth==1.1.0
Django==2.0.7
django-allauth==0.42.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
pylint==2.5.3
python3-openid==3.2.0
pytz==2018.5
redis==2.10.6
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
